- update proto support

+++ Shared/Datastructures/EntityFloatComparer.cs
@@ -10,8 +10,8 @@
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public int Compare(EntityFloat x, EntityFloat y)
         {
-            if (x.entity == 0) return 1;
-            if (y.entity == 0) return -1;
+            if ((int)x.entity == 0) return 1;
+            if ((int)y.entity == 0) return -1;
             if (x.value - y.value < 0) return -1;
             return 1;
         }
